Move css to separate location and update references accordingly Change single-class component exports to use default export by preference
