+++ src/index.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
 import App from './App';
-import './index.css';
+import './style/index.css';
 import registerServiceWorker from './registerServiceWorker';
 
 ReactDOM.render(
