+++ src/style/attribution.css
@@ -1,38 +1,5 @@
-.app {
-  display: flex;
-}
-
-.timeline-view {
-  /* border: 1px solid gray; */
-  height: 22px;
-  margin-left: 10px;
-  margin-right: 10px;
-}
-
-.ui {
-  font-size: 0.8em;
-  max-width: 300px;
-  margin-top: 10px;
-  margin-left: 20px;
-}
-
-.ui .section{
-  border: 1px solid silver;
-  padding: 10px;
-  margin-bottom: 10px;
-}
-
-.ui .sim-adjustment div span{
-  display: inline-block;
-  width: 180px;
-}
-.ui .sim-adjustment div input{
-  display: inline-block;
-  width: 80px;
-}
-
 .attribution .attribution-logo{
-  background: url('./images/cclogo.png');
+  background: url('../images/cclogo.png');
   background-size: 200px 84px;
   background-repeat: no-repeat;
   position: fixed;
@@ -82,7 +49,7 @@
 
 .attribution-text .small-logo{
 
-  background: url('./images/cclogo.png');
+  background: url('../images/cclogo.png');
   background-size: 100px 42px;
   width:100px;
   height:42px;
